% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_chapter.R
\name{single_chapter}
\alias{single_chapter}
\title{Creates pdf document from a single chapter of thesis with references}
\usage{
single_chapter(chapter, bib_file = "bib/thesis.bib")
}
\arguments{
\item{chapter}{filename of chapter to render.}

\item{bib_file}{filepath to bibtex file for references. Defaults to
"bib/thesis.bib" the default for thesisdownrli.}
}
\description{
In general the references in a thesis will be in a separate references
section but a student will want an individual chapter to be proof read
by someone else (e.g. a supervisor). This function makes this easier by
creating a temporary R markdown file adding a references section and
with the chapter to be rendered as a child document.
}
\examples{
\dontrun{
single_chapter("01-chap1.Rmd")
}
}
