% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdownrli.R
\docType{package}
\name{thesisdownrli}
\alias{thesisdownrli}
\alias{thesisdownrli-package}
\title{thesisdownrli: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
thesisdownrli: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
